Move test fixtures.
